pyproject.toml
[tool.poetry]
name = "polars-transformation-engine"
version = "0.1.0"
description = ""
authors = ["Jefferson Marchetti Ferreira <jeffersonm.ferreira@gmail.com>"]
readme = "README.md"
packages = [{include = "polars_transformation_engine"}, {include = "libs"}]

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.29.0"
asyncio = "^3.4.3"
pydantic = "^2.7.1"
polars = "^0.20.25"
isort = "^5.13.2"
confluent-kafka = "^2.4.0"
python-decouple = "^3.8"
redis = "^5.0.4"


boto3 = "^1.34.108"
deltalake = "^0.17.4"
minio = "^7.2.7"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
mypy = "^0.991"
isort = "^5.11.4"
pytest = "^7.2.1"
ipython = "^8.10.0"
pre-commit = "^3.2.1"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''